+++ layer/context.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from layer.contracts.assets import AssetType
 from layer.contracts.datasets import DatasetBuild
-from layer.tracker.progress_tracker import BaseRunProgressTracker
+from layer.tracker.ui_progress_tracker import RunProgressTracker
 from layer.training.base_train import BaseTrain
 
 
@@ -18,13 +18,13 @@ class Context:
         self,
         train: Optional[BaseTrain] = None,
         dataset_build: Optional[DatasetBuild] = None,
-        tracker: Optional[BaseRunProgressTracker] = None,
+        tracker: Optional[RunProgressTracker] = None,
         asset_name: Optional[str] = None,
         asset_type: Optional[AssetType] = None,
     ) -> None:
         self._train: Optional[BaseTrain] = train
         self._dataset_build: Optional[DatasetBuild] = dataset_build
-        self._tracker: Optional[BaseRunProgressTracker] = tracker
+        self._tracker: Optional[RunProgressTracker] = tracker
         self._asset_name = asset_name
         self._asset_type = asset_type
 
@@ -55,7 +55,7 @@ class Context:
     def with_dataset_build(self, dataset_build: Optional[DatasetBuild]) -> None:
         self._dataset_build = dataset_build
 
-    def with_tracker(self, tracker: BaseRunProgressTracker) -> None:
+    def with_tracker(self, tracker: RunProgressTracker) -> None:
         self._tracker = tracker
 
     def with_asset_name(self, asset_name: str) -> None:
@@ -64,7 +64,7 @@ class Context:
     def with_asset_type(self, asset_type: AssetType) -> None:
         self._asset_type = asset_type
 
-    def tracker(self) -> Optional[BaseRunProgressTracker]:
+    def tracker(self) -> Optional[RunProgressTracker]:
         return self._tracker
 
     def asset_name(self) -> Optional[str]:
