+++ test/e2e/test_resources.py
@@ -9,7 +9,7 @@ import layer
 from layer.contracts.projects import Project
 from layer.decorators import dataset, model, pip_requirements, resources
 from layer.resource_manager import ResourceManager
-from layer.tracker.progress_tracker import RunProgressTracker
+from layer.tracker.ui_progress_tracker import UIRunProgressTracker
 from test.e2e.assertion_utils import E2ETestAsserter
 
 
@@ -26,7 +26,7 @@ def test_resource_manager(initialized_project: Project, asserter: E2ETestAsserte
     resource_manager.wait_resource_upload(
         project_full_name,
         functions,
-        RunProgressTracker(url=URL(""), account_name="", project_name=""),
+        UIRunProgressTracker(url=URL(""), account_name="", project_name=""),
     )
     with tempfile.TemporaryDirectory(prefix="test_resource_manager") as resource_dir:
         resource_manager.wait_resource_download(
